Push down DCP1-specific method
