+++ test/indexer/test_indexer.py
@@ -11,7 +11,9 @@ from collections.abc import (
 from concurrent.futures import (
     ThreadPoolExecutor,
 )
-import copy
+from copy import (
+    deepcopy,
+)
 from itertools import (
     chain,
 )
@@ -68,6 +70,7 @@ from azul.indexer.index_service import (
     IndexExistsAndDiffersException,
     IndexService,
     IndexWriter,
+    Tallies,
     log as index_service_log,
 )
 from azul.json import (
@@ -177,6 +180,16 @@ class DCP1IndexerTestCase(DCP1TestCase, IndexerTestCase):
         actual = dict.fromkeys(expected.keys(), 0)
         actual |= Counter(self._parse_index_name(h)[1] for h in hits)
         self.assertDictEqual(expected, actual)
+
+    @classmethod
+    def _write_transforms(cls, bundle: HCABundle) -> Tallies:
+        bundle = attr.evolve(bundle,
+                             manifest=deepcopy(bundle.manifest),
+                             metadata_files=deepcopy(bundle.metadata_files))
+        transforms = cls.index_service.transform(cls.catalog, bundle, delete=False)
+        contributions, replicas = transforms
+        cls.index_service.replicate(cls.catalog, replicas)
+        return cls.index_service.contribute(cls.catalog, contributions)
 
 
 class TestDCP1Indexer(DCP1IndexerTestCase):
@@ -546,7 +559,7 @@ class TestDCP1IndexerWithIndexesSetUp(DCP1IndexerTestCase):
 
     def _patch_bundle(self, bundle: Bundle) -> str:
         new_file_uuid = str(uuid4())
-        bundle.manifest = copy.deepcopy(bundle.manifest)
+        bundle.manifest = deepcopy(bundle.manifest)
         file_name = '21935_7#154_2.fastq.gz'
         for file in bundle.manifest:
             if file['name'] == file_name:
